+++ fabric/src/main/java/com/lx862/jcm/mod/scripting/mtr/eyecandy/EyecandyBlockEntityWrapper.java
@@ -1,8 +1,6 @@
 package com.lx862.jcm.mod.scripting.mtr.eyecandy;
 
-import com.lx862.jcm.mod.util.JCMUtil;
-import org.mtr.mapping.holder.BlockPos;
-import org.mtr.mapping.holder.Vector3f;
+import com.lx862.mtrscripting.util.Vector3dWrapper;
 import org.mtr.mapping.holder.World;
 import org.mtr.mod.block.BlockEyeCandy;
 
@@ -45,12 +43,14 @@ public class EyecandyBlockEntityWrapper {
         return be.getFullBrightness();
     }
 
-    public Vector3f pos() {
-        return JCMUtil.blockPosToVector3f(be.getPos2());
+    public Vector3dWrapper pos() {
+        Vector3dWrapper realPos = new Vector3dWrapper(be.getPos2());
+        realPos.add(getTranslateX(), getTranslateY(), getTranslateZ());
+        return realPos;
     }
 
-    public BlockPos blockPos() {
-        return be.getPos2();
+    public Vector3dWrapper blockPos() {
+        return new Vector3dWrapper(be.getPos2());
     }
 
     public World getWorld() {